+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Config
     - Added support for `libs` options in model config. Libs module available in model's modules as `libs`
+- Build
+    - Fixed no favicon when favicon is not set up in a config (use default instead)
 - Client
     - Fixed output of values that contain HTML in `signature` view